refactor: clean comments and create folders to organize components. Also, includes and arrow allowing to return from the answers screen to results screen
